Changes to make targets to avoid needing to install QML plugin to run autopilot tests.
Adds a symlink named "UbuntuOne -> ." to the build dir for the QML plugin to make it loadable from the build dir.
Adds an argument to the run script to pass in the path to import the QML plugin from while running the tests.
Adds a dependency on the QML plugin from the autopilot tests.

NOTE: this works as long as the plugin is built correctly.
If something goes wrong and QML can't find the plugin, it may still find it in the installed system location (which we can't override) and you'll be testing the wrong thing.
Fortunately, the plugin can be pretty broken and QML will find it. As long as qmldir gets generated and the symlink is created, almost anything else can go wrong and you'll still be testing the right thing.

Also a clarification of an outdated comment about variables used in the qmldir template.